Modularize protocol structures
