+++ client-base/src/protocol/left.rs
@@ -1,0 +1,12 @@
+use crate::{util::SCResult, xml_node::{FromXmlNode, XmlNode}};
+
+/// A message indicating that the client
+/// has left a room with the specified id.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct Left {
+    pub room_id: String
+}
+
+impl FromXmlNode for Left {
+    fn from_node(node: &XmlNode) -> SCResult<Self> { Ok(Self { room_id: node.attribute("roomId")?.to_owned() }) }
+}
