+++ client-base/src/protocol/score_definition.rs
@@ -1,0 +1,17 @@
+use crate::{util::SCResult, xml_node::FromXmlNode, xml_node::XmlNode};
+
+use super::ScoreFragment;
+
+/// The definition of a score.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct ScoreDefinition {
+    pub fragments: Vec<ScoreFragment>
+}
+
+impl FromXmlNode for ScoreDefinition {
+    fn from_node(node: &XmlNode) -> SCResult<Self> {
+        Ok(Self {
+            fragments: node.childs_by_name("fragment").map(ScoreFragment::from_node).collect::<SCResult<_>>()?
+        })
+    }
+}
